Core: generalise capture frames to f32 with a real/complex layout
Phase 10.0 D1b: ChannelCapture.data is f32 with per-component IqCal, and
CaptureFrame carries the frame-level SampleLayout. Scope backends convert
their i8 +/-125 wire encoding at frame construction, so scope behaviour is
unchanged. .nwc is v3 (layout + IqCal; v1/v2 still read). The GPU buffer is
array<f32> with NaN gap sentinels and NO_DATA is f32::NAN.

Sim only. All tests, WGSL validation, and the windowed pixel, geometry and
flow tests pass. Budget 33 files / net 362 lines (cap recorded at 35).

+++ crates/neowon-dsp/src/decode/onewire.rs
@@ -104,25 +104,25 @@ mod tests {
     use super::*;
 
     struct Builder {
-        raw: Vec<i8>,
+        raw: Vec<f32>,
         rate: f64,
     }
 
     impl Builder {
         fn new(rate: f64) -> Self {
             Self {
-                raw: vec![100; (rate * 20e-6) as usize],
+                raw: vec![100.0; (rate * 20e-6) as usize],
                 rate,
             }
         }
         fn low(&mut self, us: f64) -> &mut Self {
             let n = (self.rate * us * 1e-6) as usize;
-            self.raw.extend(std::iter::repeat_n(-100i8, n));
+            self.raw.extend(std::iter::repeat_n(-100.0f32, n));
             self
         }
         fn high(&mut self, us: f64) -> &mut Self {
             let n = (self.rate * us * 1e-6) as usize;
-            self.raw.extend(std::iter::repeat_n(100i8, n));
+            self.raw.extend(std::iter::repeat_n(100.0f32, n));
             self
         }
         fn reset(&mut self) -> &mut Self {
